modif weight to decimal et duration to rechargement

+++ APPLICATION/Projet Finale - BD/ProjetFinale/DataAccessLayer/Factory/ItemFact.cs
@@ -17,7 +17,7 @@ namespace ProjetFinale.DataAccessLayer.Factory
             string name = reader.GetString("ObjectName");
             string diceToRoll = reader.GetString("DiceToRoll") ?? "Aucun";
             string description = reader.GetString("ObjectDescription");
-            int weight = reader.GetInt32("Weight");
+            decimal weight = reader.GetDecimal("Weight");
 
             Item newItem = new Item(id, name, diceToRoll, description, weight);
             return newItem;
